Changed Success Page to SSR
